@@ -35,16 +35,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.12",
-    "@angular/cdk": "^16.2.12",
-    "@angular/common": "^16.2.12",
-    "@angular/compiler": "^16.2.12",
-    "@angular/core": "^16.2.12",
-    "@angular/forms": "^16.2.12",
-    "@angular/material": "^16.2.12",
-    "@angular/platform-browser": "^16.2.12",
-    "@angular/platform-browser-dynamic": "^16.2.12",
-    "@angular/router": "^16.2.12",
+    "@angular/animations": "^17.3.11",
+    "@angular/cdk": "^17.3.10",
+    "@angular/common": "^17.3.11",
+    "@angular/compiler": "^17.3.11",
+    "@angular/core": "^17.3.11",
+    "@angular/forms": "^17.3.11",
+    "@angular/material": "^17.3.10",
+    "@angular/platform-browser": "^17.3.11",
+    "@angular/platform-browser-dynamic": "^17.3.11",
+    "@angular/router": "^17.3.11",
     "copyfiles": "^2.4.1",
     "csstype": "^3.1.3",
     "d3": "^7.8.5",
@@ -63,17 +63,17 @@
     "tslib": "^2.3.0",
     "xmldoc": "^1.2.0",
     "yaml": "^2.1.1",
-    "zone.js": "^0.13.3"
+    "zone.js": "^0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.11",
-    "@angular-eslint/builder": "16.3.1",
-    "@angular-eslint/eslint-plugin": "16.3.1",
-    "@angular-eslint/eslint-plugin-template": "16.3.1",
-    "@angular-eslint/schematics": "16.3.1",
-    "@angular-eslint/template-parser": "16.3.1",
-    "@angular/cli": "~16.2.11",
-    "@angular/compiler-cli": "^16.2.12",
+    "@angular-devkit/build-angular": "^17.3.8",
+    "@angular-eslint/builder": "17.5.2",
+    "@angular-eslint/eslint-plugin": "17.5.2",
+    "@angular-eslint/eslint-plugin-template": "17.5.2",
+    "@angular-eslint/schematics": "17.5.2",
+    "@angular-eslint/template-parser": "17.5.2",
+    "@angular/cli": "~17.3.8",
+    "@angular/compiler-cli": "^17.3.11",
     "@compodoc/compodoc": "^1.1.19",
     "@percy/cli": "^1.8.1",
     "@percy/cypress": "^3.1.2",
@@ -82,10 +82,10 @@
     "@types/geojson": "^7946.0.10",
     "@types/jasmine": "~4.0.0",
     "@types/topojson-client": "^3.1.1",
-    "@typescript-eslint/eslint-plugin": "^5.59.2",
-    "@typescript-eslint/parser": "^5.59.2",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
     "cypress": "^13.6.2",
-    "eslint": "^8.39.0",
+    "eslint": "^8.57.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "jasmine-core": "~4.1.0",
@@ -96,9 +96,9 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "karma-junit-reporter": "^2.0.1",
     "local-cypress": "^1.2.6",
-    "ng-packagr": "^16.2.3",
+    "ng-packagr": "^17.3.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
-    "typescript": "~4.9.5"
+    "typescript": "~5.4.5"
   }
 }
