@@ -11,8 +11,9 @@
   ],
   "schematics": "./schematics/collection.json",
   "peerDependencies": {
-    "@angular/common": ">=14.0.0",
-    "@angular/core": ">=14.0.0",
+    "@angular/cdk": "^17.0.0",
+    "@angular/common": "^17.0.0",
+    "@angular/core": "^17.0.0",
     "csstype": "^3.1.3",
     "d3": "^7.6.1",
     "file-saver": "^2.0.5",
@@ -20,14 +21,10 @@
     "lodash-es": "^4.17.21",
     "papaparse": "^5.3.2",
     "polylabel": "^1.1.0",
+    "rxjs": "^7.4.0",
     "topojson-client": "^3.1.0"
   },
   "dependencies": {
     "tslib": "^2.3.0"
-  },
-  "devDependencies": {
-    "@angular/common": "^14.0.0",
-    "@angular/core": "^14.0.0",
-    "d3": "^7.4.4"
   }
 }
