@@ -31,7 +31,7 @@ describe('the YAxis mixin', () => {
     });
     it('calls getTranslateDistance once', () => {
       abstractClass.translate$
-        .subscribe((str) => {
+        .subscribe(() => {
           expect(abstractClass.getTranslateDistance).toHaveBeenCalledOnceWith(
             testRanges
           );
